requirements.txt
pyyaml
colorama
docker
ansible